Update test to bypass scalar promotion bug

+++ tests/python_frontend/call_sdfg_with_symbols_test.py
@@ -5,18 +5,18 @@ N = dace.symbol("N")
 
 
 @dace.program
-def add_one(A: dace.int32[N, N], result: dace.int32[N, N]):
+def add_one(A: dace.int64[N, N], result: dace.int64[N, N]):
     result[:] = A + 1
 
 
 def call_test():
     @dace.program
-    def add_one_more(A: dace.int32[N, N]):
-        result = dace.define_local([N, N], dace.int32)
+    def add_one_more(A: dace.int64[N, N]):
+        result = dace.define_local([N, N], dace.int64)
         add_one(A, result)
         return result + 1
 
-    A = np.random.randint(0, 10, size=(11, 11), dtype=np.int32)
+    A = np.random.randint(0, 10, size=(11, 11), dtype=np.int64)
     result = add_one_more(A=A.copy())
     assert np.allclose(result, A + 2)
 
@@ -25,12 +25,12 @@ def call_sdfg_test():
     add_one_sdfg = add_one.to_sdfg()
 
     @dace.program
-    def add_one_more(A: dace.int32[N, N]):
-        result = dace.define_local([N, N], dace.int32)
+    def add_one_more(A: dace.int64[N, N]):
+        result = dace.define_local([N, N], dace.int64)
         add_one_sdfg(A=A, result=result)
         return result + 1
 
-    A = np.random.randint(0, 10, size=(11, 11), dtype=np.int32)
+    A = np.random.randint(0, 10, size=(11, 11), dtype=np.int64)
     result = add_one_more(A=A.copy())
     assert np.allclose(result, A + 2)
 
@@ -39,8 +39,8 @@ other_N = dace.symbol("N")
 
 
 @dace.program
-def add_one_other_n(A: dace.int32[other_N - 1, other_N - 1],
-                    result: dace.int32[other_N - 1, other_N - 1]):
+def add_one_other_n(A: dace.int64[other_N - 1, other_N - 1],
+                    result: dace.int64[other_N - 1, other_N - 1]):
     result[:] = A + 1
 
 
@@ -48,12 +48,12 @@ def call_sdfg_same_symbol_name_test():
     add_one_sdfg = add_one_other_n.to_sdfg()
 
     @dace.program
-    def add_one_more(A: dace.int32[N, N]):
-        result = dace.define_local([N, N], dace.int32)
+    def add_one_more(A: dace.int64[N, N]):
+        result = dace.define_local([N, N], dace.int64)
         add_one_sdfg(A=A, result=result)
         return result + 1
 
-    A = np.random.randint(0, 10, size=(11, 11), dtype=np.int32)
+    A = np.random.randint(0, 10, size=(11, 11), dtype=np.int64)
     result = add_one_more(A=A.copy())
     assert np.allclose(result, A + 2)
